Update bash_prompt style and git details
Make use of @cowboy's git_prompt function to display more detailed
indication of the status of the git branch.

Modify the way colors are set so that the colors can be specified
separately from the format of the prompt.

Tweaks to the prompt format. Style changes for the user and host
parts of the prompt when root or shh respectively.