Set git credentials after update
After running `bootstrap.sh` the `.gitconfig` file is replaced with the
version found in the repository. That version does not include any git
user credentials and this can cause a commit immediately after a
dotfiles update to not use the correct committer details.

To get around this problem, the `~/.extra` file is executed after the
bootstrap update process. This ensures that the `.gitconfig` file has
the correct user credentials immediately written to it.