git: update pretty log format
Move the pretty log format into a `[pretty]` entry so that it can more
easily be used by other aliases. The new pretty log format places the
author name and date at the end of the line since the sha, commit
message, and branch tend to be the more important information. The
branch is now displayed as cyan, to match the color used for the branch
display in my bash prompt.

Since the `git graph` alias is identical to `git logp` apart from the
`--graph` flag, use the `logp` alias in the `graph` alias.