bin: delete excluded files from backup volume
Any files or directories listed in the backup-excluded files will now be
deleted from the destination volume if they are present.

There may be circumstance in which this behaviour is not desired, e.g.,
for very large files (like virtual machines) that you do not want to
keep in sync with the backup volume. In those cases, remove the option
that is added by this commit and then add the files or directories to
the excludes file.