Add Homebrew script and considerations
As part of the installation and update process, Homebrew will now be
updated, formulae upgraded, and specific formulae installed if they are
not already.

This merits the creation of a new `script` directory to house the
`bootstrap` and `homebrew` scripts. There are also updates to the
`exports` bash file to make sure that Homebrew bins are added to the
path (irrespective of the location of Homebrew installations).