bash: fix to work with custom `brew` location
A bug in the previous commit meant that when Homebrew was installed in a
custom location, there were references to `brew --prefix` before `brew`
was added to the PATH.

The solution in this commit is to ask the user to add the custom `bin`
location to their PATH in `.bash_profile.local`. Then any references to
`brew --prefix` are avoided until `.bash_profile.local` is sourced in
`.bash_profile`. This avoids errors.