Replace `echo` with `printf`
`echo` is not considered portable; it differs in different shells. It's
generally recommended that `echo` is never used with options; `printf`
should be used instead as it is much more consistent across
implementations.

However, I'm going to use `printf` exclusively to avoid any potential
issues with `echo` and I prefer the consistency of using just one
command.