bin: Fix `rsync` options in backup script
Rely on options that are found in rsync 3.0 (installed via Homebrew)
in order to have the backup script work on a wider range of OS versions.

Use the `rsync` command directly so that the script doesn't have an
explicit reliance on the command mapping to rsync installed in the
`usr/bin` or Homebrew directory.