Remove PhantonJS bash export
Again, this tries to reference to `brew` command before the environment
knows the location of a custom Homebrew installation. Loading
`bash_exports` after local bash settings would defeat the point. So for
Karma to know PhantonJS's bin location when Homebrew is not in the
default directory, the bin export will need to be defined in the
`.bash_profile.local` file.