shell: use `~/.dotfilesrc` for custom `brew`
Rather than messing around with when to load `~/.bash_profile.local` and
always being mindful of whether a custom `brew` PATH has been sourced,
expect the custom path to be set in a new file: `~/.dotfilesrc`. It can
be used for any up-front configuration that is needed, while
`~/.bash_profile.local` continues to be for local Bash overrides and
settings.